Overhaul Flask implementation
